% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{squba_update}
\alias{squba_update}
\title{Update squba packages to align with most recent GitHub commits}
\usage{
squba_update()
}
\value{
a message in the console with information about which
packages need to be updated based on what is currently
installed
}
\description{
Update squba packages to align with most recent GitHub commits
}
\examples{
\dontrun{
## Requires that a GitHub PAT is established

squba_update()
}
}
